editing/empty-elements-insertion.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Placing selection and typing inside empty elements</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<style>
  #styled {
    display: inline-block;
    border: 2px red solid;
    padding: 1em;
  }
</style>
<div contenteditable>
  <p><strong id="styled"></strong></p>
</div>
<div contenteditable>
  <p>Lorem <strong id="unstyled"></strong> ipsum</p>
</div>
<script>
  promise_test( async () => {
    const target = document.querySelector( '#styled' );
    const actions = new test_driver.Actions();
    await actions
      .pointerMove( 0, 0, { origin: target } )
      .pointerDown( { button: actions.ButtonType.LEFT } )
      .pointerUp( { button: actions.ButtonType.LEFT } )
      .send();
    document.execCommand( 'insertText', false, 'a' );
    assert_greater_than( target.innerHTML.length, 0, 'The text should be inserted into the styled <strong> element' );
  }, 'Insert text into the inline element styled with border and padding' );

  promise_test( async () => {
    const target = document.querySelector( '#unstyled' );
    const actions = new test_driver.Actions();
    await actions
      .pointerMove( 0, 0, { origin: target } )
      .pointerDown( { button: actions.ButtonType.LEFT } )
      .pointerUp( { button: actions.ButtonType.LEFT } )
      .send();
    document.execCommand( 'insertText', false, 'a' );
    assert_greater_than( target.innerHTML.length, 0, 'The text should be inserted into the unstyled <strong> element' );
  }, 'Insert text into the unstyled inline element' );
</script>
